Extract FileContent to its own file.
